+++ tests/testthat/test-path_functions.R
@@ -8,6 +8,9 @@ test_that("get_abs_paths() works as expected", {
   # Project path
   expect_true(endsWith(paths$project_path, "Fellowship 1960-2015 PFU database"))
 
+  # Schema path
+  expect_true(endsWith(paths$schema_path, "Schema.xlsx"))
+
   # IEA folder path
   expect_true(endsWith(paths$iea_folder_path, file.path("IEA extended energy balance data", "IEA 2022 energy balance data")))